pyproject.toml
[project]
name = "kbiproject"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pymupdf (>=1.26.0,<2.0.0)",
    "playwright (>=1.52.0,<2.0.0)",
    "openai (>=1.84.0,<2.0.0)",
    "azure-core (>=1.34.0,<2.0.0)",
    "numpy (>=2.3.0,<3.0.0)",
    "faiss-cpu (>=1.11.0,<2.0.0)",
    "tiktoken (>=0.9.0,<0.10.0)"
]

[tool.poetry]
packages = [{include = "kbiproject", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
